+++ src/main/kotlin/io/gijung/aichatbot/auth/infrastructure/jwt/JwtGeneratorImpl.kt
@@ -2,7 +2,7 @@ package io.gijung.aichatbot.auth.infrastructure.jwt
 
 import io.gijung.aichatbot.auth.application.port.out.JwtGenerator
 import io.gijung.aichatbot.auth.domain.model.UserRole
-import io.gijung.aichatbot.user.domain.modle.vo.UserId
+import io.gijung.aichatbot.user.domain.model.vo.UserId
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.SignatureAlgorithm
 import org.springframework.stereotype.Component
