+++ src/main/kotlin/io/gijung/aichatbot/auth/presentation/web/response/SignUpResponse.kt
@@ -1,6 +1,6 @@
 package io.gijung.aichatbot.auth.presentation.web.response
 
-import io.gijung.aichatbot.user.domain.modle.vo.UserId
+import io.gijung.aichatbot.user.domain.model.vo.UserId
 
 data class SignUpResponse(
     val id: String
