+++ src/test/kotlin/io/gijung/aichatbot/auth/application/service/MockitoHelper.kt
@@ -1,6 +1,6 @@
 package io.gijung.aichatbot.auth.application.service
 
-import io.gijung.aichatbot.user.domain.modle.vo.UserId
+import io.gijung.aichatbot.user.domain.model.vo.UserId
 import org.mockito.Mockito
 
 object MockitoHelper {
